Delete folder errors

+++ src/services/files.ts
@@ -80,7 +80,13 @@ export async function copyFolder(sourcePath: string, targetPath: string) {
 }
 
 export function deleteFolder(folderPath: string) {
-  fs.removeSync(folderPath);
+  if (fs.existsSync(folderPath)) {
+    fs.removeSync(folderPath);
+  } else {
+    throw new Error(
+      `Trying to delete a non-existing directory: "${folderPath}"`,
+    );
+  }
 }
 
 export async function unzipRemoteFile(
